+++ src/inventorum.ebay/inventorum/ebay/tests/utils.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
 from decimal import Decimal
+import json
 import logging
+import os
+from inventorum.ebay import settings
 import mock
 
 
@@ -28,3 +31,15 @@ class AssertionMixin(object):
         first = Decimal(str(first))
         second = Decimal(str(second))
         self.assertEqual(first, second, msg)
+
+
+class JSONFixture(object):
+
+    @classmethod
+    def load(cls, path):
+        file_path = os.path.join(settings.FIXTURE_DIR, path)
+
+        with open(file_path) as data:
+            data = json.load(data)
+
+        return data